perf: optimise dashboard illustration + separate chart components from index
