+++ src/pages/Market/index.stories.tsx
@@ -1,11 +1,11 @@
 import { Meta } from '@storybook/react';
 import BigNumber from 'bignumber.js';
-import { ApyChartProps } from 'components';
 import React from 'react';
 
 import { assetData } from '__mocks__/models/asset';
 import { poolData } from '__mocks__/models/pools';
 import { vTokenApySimulations } from '__mocks__/models/vTokenApySimulations';
+import { type ApyChartProps } from 'components/charts/ApyChart';
 import { withRouter } from 'stories/decorators';
 
 import { MarketUi } from '.';
